Fixed a bug in histograms for windows

+++ Histograms/src/Histograms.tpp
@@ -540,7 +540,7 @@ class cdigitizer_histogram2D_steps
 			count = 0;
             #ifdef _WIN32_WINNT
                 uint64_t nbgroups = GetActiveProcessorGroupCount();
-                N_t = std::min(64,omp_get_max_threads()*nbgroups);
+                N_t = std::min((uint64_t) 64,omp_get_max_threads()*nbgroups);
             #else
                 N_t = omp_get_max_threads();
 			#endif
